A set of video layout enhancements (#58)
* Smooth the transition from camera-tab video panel to external video panel

* Fix issue with saving empty video layout

* Added close video-layout button and an image for the delete-layout button

* Added ctrl button option to open additional layouts

* Smooth the transition from old-layout video panels to new-layout video panels

* Add confirmation-dialogs to save-layout and delete-layout operations.

+++ src/us/mn/state/dot/tms/client/camera/StreamControlPanel.java
@@ -2,6 +2,7 @@
  * IRIS -- Intelligent Roadway Information System
  * Copyright (C) 2002-2020  Minnesota Department of Transportation
  * Copyright (C) 2014-2015  AHMCT, University of California
+ * Copyright (C) 2022-2023  SRF Consulting Group
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,6 +17,7 @@
 package us.mn.state.dot.tms.client.camera;
 
 import java.awt.FlowLayout;
+import java.awt.Frame;
 import java.awt.event.ActionEvent;
 import java.util.ArrayList;
 import java.util.Properties;
@@ -24,7 +26,11 @@ import javax.swing.DefaultComboBoxModel;
 import javax.swing.ImageIcon;
 import javax.swing.JButton;
 import javax.swing.JComboBox;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
+import javax.swing.SwingUtilities;
+import javax.swing.SwingWorker;
+
 import us.mn.state.dot.tms.client.EditModeListener;
 import us.mn.state.dot.tms.client.Session;
 import us.mn.state.dot.tms.client.UserProperty;
@@ -42,6 +48,7 @@ import static us.mn.state.dot.tms.client.widget.Widgets.UI;
  * @author Travis Swanston
  * @author Michael Janson
  * @author Gordon Parikh
+ * @author John L. Stanley
  */
 public class StreamControlPanel extends JPanel {
 
@@ -50,9 +57,10 @@ public class StreamControlPanel extends JPanel {
 		STOP("camera.stream.stop"),
 		PLAY("camera.stream.play"),
 		PLAY_EXTERNAL("camera.stream.playext"),
+		RESTORE_LAYOUT("camera.layout.restore"),
+		CLOSE_ALL_LAYOUTS("camera.layout.close"),
 		SAVE_LAYOUT("camera.layout.save"),
-		DELETE_LAYOUT("camera.layout.delete"),
-		RESTORE_LAYOUT("camera.layout.restore");
+		DELETE_LAYOUT("camera.layout.delete");
 
 		/** Command I18n text */
 		private final String text_id;
@@ -67,7 +75,7 @@ public class StreamControlPanel extends JPanel {
 			IAction ia = new IAction(text_id) {
 				@Override
 				protected void doActionPerformed(ActionEvent ev) {
-					handleButton(pnl);
+					handleButton(pnl, ev);
 				}
 			};
 			final JButton btn = new JButton(ia);
@@ -81,8 +89,9 @@ public class StreamControlPanel extends JPanel {
 			return btn;
 		}
 
-		/** Handle control button press */
-		private void handleButton(StreamControlPanel pnl) {
+		/** Handle button press */
+		private void handleButton(StreamControlPanel pnl, ActionEvent ev) {
+			boolean bCtrl = (ev.getModifiers() & ActionEvent.CTRL_MASK) != 0;
 			switch (this) {
 			case STOP:
 				pnl.stopStream();
@@ -93,14 +102,17 @@ public class StreamControlPanel extends JPanel {
 			case PLAY_EXTERNAL:
 				pnl.playExternal();
 				break;
+			case RESTORE_LAYOUT:
+				pnl.restoreLayout(bCtrl);
+				break;
+			case CLOSE_ALL_LAYOUTS:
+				pnl.closeAllLayouts();
+				break;
 			case SAVE_LAYOUT:
 				pnl.saveLayout();
 				break;
 			case DELETE_LAYOUT:
 				pnl.deleteLayout();
-				break;
-			case RESTORE_LAYOUT:
-				pnl.restoreLayout();
 				break;
 			}
 		}
@@ -149,6 +161,9 @@ public class StreamControlPanel extends JPanel {
 	/** Restore layout button */
 	private final JButton restore_layout_btn;
 
+	/** Close all layouts button */
+	private final JButton close_all_layouts_btn;
+
 	/** Create a stream control panel.
 	 * @param s User session. */
 	public StreamControlPanel(Session s, StreamPanel pnl) {
@@ -165,20 +180,24 @@ public class StreamControlPanel extends JPanel {
 		layout_cbx = new JComboBox<String>(layout_mdl);
 		layout_cbx.setToolTipText(I18N.get("camera.layout"));
 		layout_cbx.setPrototypeDisplayValue("layout_MMM");
+		restore_layout_btn = StreamCommand.RESTORE_LAYOUT
+				.createButton(this);
+		close_all_layouts_btn = StreamCommand.CLOSE_ALL_LAYOUTS
+				.createButton(this);
 		save_layout_btn = StreamCommand.SAVE_LAYOUT.createButton(this);
 		delete_layout_btn = StreamCommand.DELETE_LAYOUT
-			.createButton(this);
-		restore_layout_btn = StreamCommand.RESTORE_LAYOUT
-			.createButton(this);
-
+				.createButton(this);
+		
 		add(stop_btn);
 		add(play_btn);
 		add(playext_btn);
-		add(Box.createHorizontalStrut(UI.hgap));
+		add(Box.createHorizontalStrut(UI.hgap*2));
 		add(layout_cbx);
+		add(restore_layout_btn);
+		add(close_all_layouts_btn);
+		add(Box.createHorizontalStrut(UI.hgap*2));
 		add(save_layout_btn);
 		add(delete_layout_btn);
-		add(restore_layout_btn);
 
 		updateButtonState(false, false);
 		updateLayoutList();
@@ -206,19 +225,91 @@ public class StreamControlPanel extends JPanel {
 
 	/** Save the current layout */
 	private void saveLayout() {
+		String layoutName = getLayoutName();
+		Integer num = UserProperty.getLayoutNumber(props, layoutName);
+		if (num != null) {
+			String title = I18N.get("camera.layout.save.query.title");
+			String msg   = I18N.get("camera.layout.save.query.msg");
+			msg = String.format(msg, layoutName);
+			int result = JOptionPane.showConfirmDialog(this,
+					msg, title,
+					JOptionPane.YES_NO_OPTION,
+					JOptionPane.QUESTION_MESSAGE);
+			if (result != JOptionPane.YES_OPTION)
+				return;
+		}
 		UserProperty.saveStreamLayout(props, getLayoutName());
 		updateLayoutList();
 	}
 
-	/** Restore the saved layout */
-	private void restoreLayout() {
+	/** Restore the selected layout 
+	 * @param bCtrl If true, don't auto-close previous open layout(s) */
+	private void restoreLayout(boolean bCtrl) {
 		StreamLayout layout = new StreamLayout(props, getLayoutName());
+		if (!bCtrl)
+			queueCloseOldLayouts(StreamLayout.getOpenFramesList());
 		layout.restoreFrames(desktop);
 		updateLayoutList();
+		updateCloseAllLayoutsBtn();
 	}
 
+	/** Close all open video layouts */
+	public void closeAllLayouts() {
+		ArrayList<Frame> oldFrames = StreamLayout.getOpenFramesList();
+		for (Frame f: oldFrames)
+			f.dispose();
+		updateCloseAllLayoutsBtn();
+	}
+
+	/** Queue a close old-frames to be done
+	 *   -after- all new frames are drawn.
+	 * Done this way to smooth the transition from old-
+	 * layout video panels to new-layout video panels.
+	 * (Reduces the odds that the video proxy will
+	 *  totally drop and restart a video stream.)
+	 * @param oldFrames Old layout video frames to be closed */ 
+	public void queueCloseOldLayouts(ArrayList<Frame> oldFrames) {
+		new SwingWorker<Void, Void>() {
+			@Override
+			protected Void doInBackground() throws Exception {
+				// First, wait 2sec for the new video frames to start being drawn.
+				Thread.sleep(2000);
+				return null;
+			}
+			@Override
+			protected void done() {
+				// Then, wait for all new video frames to finish being drawn.
+				SwingUtilities.invokeLater(new Runnable() 
+				{
+					public void run()
+					{
+						// Finally, close all old video frames.
+						for (Frame f: oldFrames)
+							f.dispose();
+						updateCloseAllLayoutsBtn();
+					}
+				});
+			}
+		}.execute();
+	}
+	
 	/** Delete the selected layout */
 	private void deleteLayout() {
+		Object[] options = {"Yes", "No"};
+		String layoutName = getLayoutName();
+		Integer num = UserProperty.getLayoutNumber(props, layoutName);
+		if (num != null) {
+			String title = I18N.get("camera.layout.delete.query.title");
+			String msg   = I18N.get("camera.layout.delete.query.msg");
+			msg = String.format(msg, layoutName);
+			int result = JOptionPane.showOptionDialog(this,
+					msg, title,
+					JOptionPane.YES_NO_OPTION,
+					JOptionPane.WARNING_MESSAGE,
+					null, options, options[1]);
+			if (result != JOptionPane.YES_OPTION)
+				return;
+		}
 		UserProperty.deleteStreamLayout(props, getLayoutName());
 		updateLayoutList();
 	}
@@ -240,6 +331,7 @@ public class StreamControlPanel extends JPanel {
 		for (String ln: layoutNames)
 			layout_mdl.addElement(ln);
 		layout_cbx.setSelectedItem(layoutName);
+		updateCloseAllLayoutsBtn();
 	}
 
 	/** Update the button state */
@@ -247,6 +339,7 @@ public class StreamControlPanel extends JPanel {
 		stop_btn.setEnabled(has_camera && is_streaming);
 		play_btn.setEnabled(has_camera && !is_streaming);
 		playext_btn.setEnabled(has_camera);
+		updateCloseAllLayoutsBtn();
 	}
 
 	/** Update the edit mode */
@@ -255,5 +348,10 @@ public class StreamControlPanel extends JPanel {
 		save_layout_btn.setEnabled(editMode);
 		delete_layout_btn.setEnabled(editMode);
 		layout_cbx.setEditable(editMode);
+	}
+
+	/** Update the close all layouts button */
+	private void updateCloseAllLayoutsBtn() {
+		close_all_layouts_btn.setEnabled(StreamLayout.framesAreOpen());
 	}
 }
